add metadata to templates page
